+++ src/Data/ByteString/Strict/Lens.hs
@@ -17,7 +17,6 @@ module Data.ByteString.Strict.Lens
 import Control.Lens
 import Data.ByteString as Words
 import Data.ByteString.Char8 as Char8
-import Data.List.Lens
 import Data.Word (Word8)
 
 -- | 'Data.ByteString.pack' (or 'Data.ByteString.unpack') a list of bytes into a 'ByteString'
@@ -31,11 +30,11 @@ packedBytes = iso Words.pack Words.unpack
 
 -- | Traverse each 'Word8' in a 'ByteString'
 --
--- @'bytes' = 'from' 'packedBytes' '.>' 'traverseList'@
+-- @'bytes' = 'from' 'packedBytes' '.>' 'itraversed'@
 --
 -- @'anyOf' 'bytes' ('==' 0x80) :: 'ByteString' -> 'Bool'@
 bytes :: SimpleIndexedTraversal Int ByteString Word8
-bytes = from packedBytes .> traverseList
+bytes = from packedBytes .> itraversed
 {-# INLINE bytes #-}
 
 -- | 'Data.ByteString.Char8.pack' (or 'Data.ByteString.Char8.unpack') a list of characters into a 'ByteString'
@@ -59,5 +58,5 @@ packedChars = iso Char8.pack Char8.unpack
 --
 -- @'anyOf' 'chars' ('==' \'c\') :: 'ByteString' -> 'Bool'@
 chars :: SimpleIndexedTraversal Int ByteString Char
-chars = from packedChars .> traverseList
+chars = from packedChars .> itraversed
 {-# INLINE chars #-}