+++ CHANGELOG.markdown
@@ -3,6 +3,11 @@
 * Redefined `simple` and moved it to `Control.Lens.Iso`. Instead of using `simple l` you can now compose `l.simple` or `simple.l` providing more nuanced control and a more compositional API.
 * Moved the various `foo#` combinators used to emit cleaner core into an unexported module, `Control.Lens.Unsafe`. This removes `MagicHash` from the public API.
 * Removed the `bazaar#` and `runBazaar#` coercions that caused issues on GHC HEAD.
+* Changed the default definition of `plate` to `uniplate` from `ignored`.
+* Added `Data.Vector.Lens` and instances for `Data.Vector`.
+* Removed redundant `Data.List.traverseList`. Use `itraversed` or `traverse` instead.
+* Moved `(:<->)` to `Control.Lens.Simple`.
+* Fixed a bug in `Control.Lens.TH` that was causing `makeIso` not to work.
 
 3.2
 ---